requirements.txt
flask
requests
gunicorn
tensorflow
numpy
pillow